requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
requests
asyncio
langchain 
langchain-openai
python-multipart
python-dotenv